Refactor bad practices
Refactoring of different parts of code that were repetitive and could be isolated into single modules in the form of global functionalities
