--- src/components/ordersByCategoryPlot/OrdersByCategoryPlot.scss
@@ -1,57 +1,0 @@
-@import "../../scss/variables";
-@import "../../scss/responsive";
-
-.orders-category-plot-container {
-  gap: 1rem;
-  height: 25rem;
-  border-radius: 1rem;
-  justify-content: center;
-  background-color: $dark-blue;
-  flex-direction: column;
-  transition: 0.5s ease;
-  align-items: center;
-  display: flex;
-  flex-grow: 1;
-
-  &:hover {
-    box-shadow: 8px 8px 0px 0px $pink;
-  }
-
-  @include responsive(portatil-standard) {
-    width: inherit;
-  }
-
-  @include responsive(mobile-large) {
-    padding: 1rem 0rem 1rem 0rem;
-  }
-
-  &--title {
-    font-size: 1.5rem;
-
-    @include responsive(mobile-large) {
-      width: 90%;
-    }
-  }
-}
-
-.custom-tooltip {
-  width: 8rem;
-  height: 4rem;
-  justify-content: center;
-  background-color: $soft-blue;
-  flex-direction: column;
-  border-radius: 1rem;
-  display: flex;
-  opacity: 0.9;
-
-  &--title {
-    font-size: 0.8rem;
-    margin: 0rem 0.6rem 0rem 0.6rem;
-    opacity: 0.5;
-  }
-
-  &--value {
-    font-size: 1rem;
-    margin: 0rem 0.6rem 0rem 0.6rem;
-  }
-}
