--- src/components/salesKPI/SalesKPI.scss
@@ -1,36 +1,0 @@
-@import "../../scss/variables";
-
-.sales-kpi-container {
-  gap: 0.5rem;
-  height: 10rem;
-  position: relative;
-  border-radius: 1rem;
-  justify-content: center;
-  background-color: $dark-blue;
-  flex-direction: column;
-  transition: 0.5s ease;
-  display: flex;
-  flex-grow: 1;
-
-  &:hover {
-    box-shadow: 8px 8px 0px 0px $pink;
-  }
-
-  &--title {
-    font-size: 1.5rem;
-    margin: 0rem 2rem 0rem 2rem;
-  }
-
-  &--value {
-    color: $pink;
-    margin: 0rem 2rem 0rem 2rem;
-    font-size: 2.2rem;
-  }
-
-  &--unit {
-    bottom: 0;
-    margin: 0rem 1rem 1rem 0rem;
-    position: absolute;
-    right: 0;
-  }
-}
