+++ src/components/utilities/kpiStyles.scss
@@ -1,6 +1,6 @@
 @import "../../scss/variables";
 
-.installments-kpi-container {
+.kpi-container {
   gap: 0.5rem;
   height: 10rem;
   position: relative;
@@ -29,7 +29,6 @@
 
   &--unit {
     bottom: 0;
-    font-size: 1.2rem;
     margin: 0rem 1rem 1rem 0rem;
     position: absolute;
     right: 0;
