+++ src/components/utilities/plotStyles.scss
@@ -1,7 +1,7 @@
 @import "../../scss/variables";
 @import "../../scss/responsive";
 
-.order-status-plot-container {
+.plot-container {
   gap: 1rem;
   height: 25rem;
   border-radius: 1rem;
